requirements.txt
streamlit 
newspaper3k 
torch 
transformers
lxml_html_clean
validators
